IP and cert setup consistency check
Make sure we consistency on the server side for our certificates.
We expect 1 IPv4 address and 1 IPv6 address for each there should be a
certificate. This way we have a consistent setup of all update infrastructure
servers. Also check the expiration date to ensure that we have a sufficient
amount of time to rotate the certs, which is especially important as
server certs for region servers are backed into the client via an RPM
package and that needs to go through an update cycle when server certs are
updated.